% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercicio_01.R
\name{colar}
\alias{colar}
\title{O primeiro pacote a gente nunca esquece...}
\usage{
colar(x, y)
}
\arguments{
\item{x}{Primeira String}

\item{y}{Segunda String}
}
\value{
A concatenacao das duas Strings
}
\description{
Concatena duas strings com um espaco em branco entre elas
}
