% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encontrar_pior_ano_time.R
\name{encontrar_pior_ano_time}
\alias{encontrar_pior_ano_time}
\title{Pior time}
\usage{
encontrar_pior_ano_time(time)
}
\arguments{
\item{time}{nome do time}
}
\value{
uma linha
}
\description{
Esse programa retorna o prior time da temporada. teste2 outro
}
