% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fazer_grafico.R
\name{fazer_grafico}
\alias{fazer_grafico}
\title{Faz um gráfico}
\usage{
fazer_grafico(dados, coluna)
}
\arguments{
\item{dados}{base de dados}

\item{coluna}{coluna que eu quero visualizar}
}
\description{
Faz um grafico usando ggplot2
}
