% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{olist_products}
\alias{olist_products}
\title{Dados da olist}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 32951 rows and 9 columns.
}
\usage{
olist_products
}
\description{
Dados da olist
}
\keyword{datasets}
